fase_2/tech_challenge/toggle-master-microservices/analytics-service/requirements.txt
Flask==3.0.3
gunicorn==20.1.0
python-dotenv==0.21.0
boto3==1.26.50